mueve endpoints del totem a routes agenda
